+++ Python/rectangle.py
