--- tile.py
@@ -1,4 +1,0 @@
-#Our tiles can be loaded with different sprites
-class Tile(object):
-	def __init__(self, image_file):
-		self.file = image_file
